refactor(agent): move the settings-aggregate module into the agent package
aleph.vm.utils.aggregate was made a neutral shared module in the split
endgame because the pool imported update_aggregate_settings. With the GPU
annotation moved agent-side, every consumer is in aleph.vm.agent, so the
module moves there: the existing import-linter contract (the supervisor
side does not import the agent) now statically guarantees the supervisor
never talks to the CCN API.

CompatibleGPU moves along: it types the aggregate's compatible_gpus
entries, not host hardware.
